Move StubEventBridgeClient to testing utils
