+++ testing-utils/src/main/java/no/sikt/nva/data/report/testing/utils/stubs/StubEventBridgeClient.java
@@ -1,0 +1,42 @@
+package no.sikt.nva.data.report.testing.utils.stubs;
+
+import static no.unit.nva.commons.json.JsonUtils.dtoObjectMapper;
+import static nva.commons.core.attempt.Try.attempt;
+import commons.handlers.KeyBatchRequestEvent;
+import nva.commons.core.SingletonCollector;
+import software.amazon.awssdk.services.eventbridge.EventBridgeClient;
+import software.amazon.awssdk.services.eventbridge.model.PutEventsRequest;
+import software.amazon.awssdk.services.eventbridge.model.PutEventsRequestEntry;
+import software.amazon.awssdk.services.eventbridge.model.PutEventsResponse;
+
+public class StubEventBridgeClient implements EventBridgeClient {
+
+        private KeyBatchRequestEvent latestEvent;
+
+        public KeyBatchRequestEvent getLatestEvent() {
+            return latestEvent;
+        }
+
+        public PutEventsResponse putEvents(PutEventsRequest putEventsRequest) {
+            this.latestEvent = saveContainedEvent(putEventsRequest);
+            return PutEventsResponse.builder().failedEntryCount(0).build();
+        }
+
+        @Override
+        public String serviceName() {
+            return null;
+        }
+
+        @Override
+        public void close() {
+
+        }
+
+        private KeyBatchRequestEvent saveContainedEvent(PutEventsRequest putEventsRequest) {
+            PutEventsRequestEntry eventEntry = putEventsRequest.entries()
+                                                   .stream()
+                                                   .collect(SingletonCollector.collect());
+            return attempt(eventEntry::detail).map(
+                jsonString -> dtoObjectMapper.readValue(jsonString, KeyBatchRequestEvent.class)).orElseThrow();
+        }
+    }
